+++ src/Concerns/ImageHasDefault.php
@@ -4,24 +4,12 @@ declare(strict_types=1);
 
 namespace Gravatar\Concerns;
 
+use Gravatar\Enum\DefaultImage;
 use Gravatar\Exception\InvalidDefaultImageException;
 
 trait ImageHasDefault
 {
     use ImageForceDefault, ImageHasInitials;
-
-    private const VALID_DEFAULT_IMAGES = [
-        'initials',
-        'color',
-        '404',
-        'mp',
-        'identicon',
-        'monsterid',
-        'wavatar',
-        'retro',
-        'robohash',
-        'blank',
-    ];
 
     /**
      * The default image to use ; either a string of the gravatar recognized default image "type" to use, or a URL
@@ -31,11 +19,11 @@ trait ImageHasDefault
             if ($value !== null) {
                 $value = strtolower($value);
 
-                if (! \in_array($value, self::VALID_DEFAULT_IMAGES) && ! filter_var($value, FILTER_VALIDATE_URL)) {
+                if (! \in_array($value, DefaultImage::values()) && ! filter_var($value, FILTER_VALIDATE_URL)) {
                     throw new InvalidDefaultImageException(\sprintf(
                         'The default image "%s" is not a recognized gravatar "default" and is not a valid URL, default gravatar can be: %s',
                         $value,
-                        implode(', ', self::VALID_DEFAULT_IMAGES)
+                        implode(', ', DefaultImage::values())
                     ));
                 }
             }
@@ -46,11 +34,11 @@ trait ImageHasDefault
     /**
      * Get or set the default image to use for avatars.
      *
-     * @param  string|null  $defaultImage  The default image to use. Use a valid image URL, or a recognized gravatar "default".
+     * @param  DefaultImage|string|null  $defaultImage  The default image to use. Use a valid image URL, or a recognized gravatar "default".
      * @param  bool  $forceDefault  Force the default image to be always load.
      * @return $this|string|null
      */
-    public function defaultImage(?string $defaultImage = null, bool $forceDefault = false): static|string|null
+    public function defaultImage(DefaultImage|string|null $defaultImage = null, bool $forceDefault = false): static|string|null
     {
         if ($defaultImage === null) {
             return $this->defaultImage;
@@ -62,11 +50,11 @@ trait ImageHasDefault
     /**
      * Alias for the "defaultImage" method.
      *
-     * @param  string|null  $defaultImage  The default image to use. Use a valid image URL, or a recognized gravatar "default".
+     * @param  DefaultImage|string|null  $defaultImage  The default image to use. Use a valid image URL, or a recognized gravatar "default".
      * @param  bool  $forceDefault  Force the default image to be always load.
      * @return $this|string|null
      */
-    public function d(?string $defaultImage = null, bool $forceDefault = false): static|string|null
+    public function d(DefaultImage|string|null $defaultImage = null, bool $forceDefault = false): static|string|null
     {
         return $this->defaultImage($defaultImage, $forceDefault);
     }
@@ -74,20 +62,20 @@ trait ImageHasDefault
     /**
      * Set the default image to use for avatars.
      *
-     * @param  string|null  $defaultImage  The default image to use. Use a valid image URL, or a recognized gravatar "default".
+     * @param  DefaultImage|string|null  $defaultImage  The default image to use. Use a valid image URL, or a recognized gravatar "default".
      * @param  bool  $forceDefault  Force the default image to be always load.
      * @return $this The current Gravatar Image instance.
      *
      * @throws InvalidDefaultImageException
      */
-    public function setDefaultImage(?string $defaultImage = null, bool $forceDefault = false): static
+    public function setDefaultImage(DefaultImage|string|null $defaultImage = null, bool $forceDefault = false): static
     {
         if ($forceDefault) {
             $this->enableForceDefault();
         }
 
         if ($defaultImage !== null) {
-            $this->defaultImage = $defaultImage;
+            $this->defaultImage = $defaultImage instanceof DefaultImage ? $defaultImage->value : $defaultImage;
         }
 
         return $this;