+++ src/Concerns/ImageHasExtension.php
@@ -4,12 +4,11 @@ declare(strict_types=1);
 
 namespace Gravatar\Concerns;
 
+use Gravatar\Enum\Extension;
 use Gravatar\Exception\InvalidImageExtensionException;
 
 trait ImageHasExtension
 {
-    private const VALID_EXTENSIONS = ['jpg', 'jpeg', 'gif', 'png', 'webp'];
-
     /**
      * The extension to append to the avatars URL.
      */
@@ -18,11 +17,11 @@ trait ImageHasExtension
             if ($value !== null) {
                 $value = strtolower($value);
 
-                if (! \in_array($value, self::VALID_EXTENSIONS)) {
+                if (! \in_array($value, Extension::values())) {
                     throw new InvalidImageExtensionException(\sprintf(
                         'The extension "%s" is not a valid one, extension image for Gravatar can be: "%s"',
                         $value,
-                        implode('", "', self::VALID_EXTENSIONS)
+                        implode('", "', Extension::values())
                     ));
                 }
             }
@@ -33,10 +32,10 @@ trait ImageHasExtension
     /**
      * Get or set the avatar extension to use.
      *
-     * @param  string|null  $extension  The avatar extension to use.
+     * @param  Extension|string|null  $extension  The avatar extension to use.
      * @return $this|string|null
      */
-    public function extension(?string $extension = null): static|string|null
+    public function extension(Extension|string|null $extension = null): static|string|null
     {
         if ($extension === null) {
             return $this->extension;
@@ -48,9 +47,10 @@ trait ImageHasExtension
     /**
      * Alias for the "extension" method.
      *
+     * @param  Extension|string|null  $extension  The avatar extension to use.
      * @return $this|string|null
      */
-    public function e(?string $extension = null): static|string|null
+    public function e(Extension|string|null $extension = null): static|string|null
     {
         return $this->extension($extension);
     }
@@ -58,15 +58,15 @@ trait ImageHasExtension
     /**
      * Set the avatar extension to use.
      *
-     * @param  string|null  $extension  The avatar extension to use.
+     * @param  Extension|string|null  $extension  The avatar extension to use.
      * @return $this The current Gravatar Image instance.
      *
      * @throws InvalidImageExtensionException
      */
-    public function setExtension(?string $extension = null): static
+    public function setExtension(Extension|string|null $extension = null): static
     {
         if ($extension !== null) {
-            $this->extension = $extension;
+            $this->extension = $extension instanceof Extension ? $extension->value : $extension;
         }
 
         return $this;