+++ src/Concerns/ImageHasMaxRating.php
@@ -4,12 +4,11 @@ declare(strict_types=1);
 
 namespace Gravatar\Concerns;
 
+use Gravatar\Enum\Rating;
 use Gravatar\Exception\InvalidMaxRatingImageException;
 
 trait ImageHasMaxRating
 {
-    private const VALID_MAX_RATINGS = ['g', 'pg', 'r', 'x'];
-
     /**
      * The maximum rating to allow for the avatars.
      */
@@ -18,11 +17,11 @@ trait ImageHasMaxRating
             if ($value !== null) {
                 $value = strtolower($value);
 
-                if (! \in_array($value, self::VALID_MAX_RATINGS)) {
+                if (! \in_array($value, Rating::values())) {
                     throw new InvalidMaxRatingImageException(\sprintf(
                         'Invalid rating "%s" specified, only allowed to be used are: "%s"',
                         $value,
-                        implode('", "', self::VALID_MAX_RATINGS)
+                        implode('", "', Rating::values())
                     ));
                 }
             }
@@ -33,9 +32,10 @@ trait ImageHasMaxRating
     /**
      * Get or set the maximum allowed rating for avatars.
      *
+     * @param  Rating|string|null  $maxRating  The maximum rating to use for avatars.
      * @return $this|string|null
      */
-    public function maxRating(?string $maxRating = null): static|string|null
+    public function maxRating(Rating|string|null $maxRating = null): static|string|null
     {
         if ($maxRating === null) {
             return $this->maxRating;
@@ -47,9 +47,10 @@ trait ImageHasMaxRating
     /**
      * Alias for the "rating" method.
      *
+     * @param  Rating|string|null  $maxRating  The maximum rating to use for avatars.
      * @return $this|string|null
      */
-    public function r(?string $maxRating = null): static|string|null
+    public function r(Rating|string|null $maxRating = null): static|string|null
     {
         return $this->maxRating($maxRating);
     }
@@ -57,15 +58,15 @@ trait ImageHasMaxRating
     /**
      * Set the maximum allowed rating for avatars.
      *
-     * @param  string|null  $maxRating  The maximum rating to use for avatars.
+     * @param  Rating|string|null  $maxRating  The maximum rating to use for avatars.
      * @return $this The current Gravatar Image instance.
      *
      * @throws InvalidMaxRatingImageException
      */
-    public function setMaxRating(?string $maxRating = null): static
+    public function setMaxRating(Rating|string|null $maxRating = null): static
     {
         if ($maxRating !== null) {
-            $this->maxRating = $maxRating;
+            $this->maxRating = $maxRating instanceof Rating ? $maxRating->value : $maxRating;
         }
 
         return $this;