+++ src/Concerns/ProfileHasFormat.php
@@ -4,22 +4,21 @@ declare(strict_types=1);
 
 namespace Gravatar\Concerns;
 
+use Gravatar\Enum\ProfileFormat;
 use Gravatar\Exception\InvalidProfileFormatException;
 
 trait ProfileHasFormat
 {
-    private const VALID_FORMATS = ['json', 'xml', 'php', 'vcf', 'qr'];
-
     /**
      * The format to append to the profile URL.
      */
     public private(set) ?string $format = null {
         set {
-            if ($value !== null && ! \in_array($value, self::VALID_FORMATS)) {
+            if ($value !== null && ! \in_array($value, ProfileFormat::values())) {
                 throw new InvalidProfileFormatException(\sprintf(
                     'The format "%s" is not a valid one, profile format for Gravatar can be: "%s"',
                     $value,
-                    implode('", "', self::VALID_FORMATS)
+                    implode('", "', ProfileFormat::values())
                 ));
             }
             $this->format = $value;
@@ -29,10 +28,10 @@ trait ProfileHasFormat
     /**
      * Get or set the profile format to use.
      *
-     * @param  string|null  $format  The profile format to use.
+     * @param  ProfileFormat|string|null  $format  The profile format to use.
      * @return $this|string|null
      */
-    public function format(?string $format = null): static|string|null
+    public function format(ProfileFormat|string|null $format = null): static|string|null
     {
         if ($format === null) {
             return $this->format;
@@ -44,9 +43,10 @@ trait ProfileHasFormat
     /**
      * Alias for the "format" method.
      *
+     * @param  ProfileFormat|string|null  $format  The profile format to use.
      * @return $this|string|null
      */
-    public function f(?string $format = null): static|string|null
+    public function f(ProfileFormat|string|null $format = null): static|string|null
     {
         return $this->format($format);
     }
@@ -54,15 +54,15 @@ trait ProfileHasFormat
     /**
      * Set the profile format to use.
      *
-     * @param  string|null  $format  The profile format to use
+     * @param  ProfileFormat|string|null  $format  The profile format to use
      * @return $this The current Profile instance
      *
      * @throws InvalidProfileFormatException
      */
-    public function setFormat(?string $format = null): static
+    public function setFormat(ProfileFormat|string|null $format = null): static
     {
         if ($format !== null) {
-            $this->format = $format;
+            $this->format = $format instanceof ProfileFormat ? $format->value : $format;
         }
 
         return $this;